Split RenameRefactoringAction into a part that requires suspension and one that does not
This allows to reuse the part not requiring suspension inside other refactoring actions and in refactoring preview providers. There is some code duplication between the two parts, but I chose not to split them until a third version arises.
