+++ RubberduckTests/Refactoring/Rename/RenameTests.cs
@@ -3369,7 +3369,9 @@ End Property";
             uiDispatcherMock
                 .Setup(m => m.Invoke(It.IsAny<Action>()))
                 .Callback((Action action) => action.Invoke());
-            var baseRefactoring = new RenameRefactoringAction(state, state?.ProjectsProvider, state, rewritingManager);
+            var componentRename = new RenameComponentOrProjectRefactoringAction(state, state?.ProjectsProvider, state, rewritingManager);
+            var otherRename = new RenameCodeDefinedIdentifierRefactoringAction(state, state?.ProjectsProvider, rewritingManager);
+            var baseRefactoring = new RenameRefactoringAction(componentRename, otherRename);
             return new RenameRefactoring(baseRefactoring, factory, state, state?.ProjectsProvider, selectionService, selectedDeclarationService, uiDispatcherMock.Object);
         }
